+++ flutter_plugin_qpos/CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.1.2+2
+## 0.1.3
 
-* 1.Add TextEditingController for mifare address
+* 1.update SDK version to 4.0.5(Add the flag isOperateMifare to make mifare classic support write value without conversion)
+* 2.Add TextEditingController for mifare value
+* 3.Add resetQPosStatus method for ios